Documentation done

+++ include/public/api.Task.hpp
@@ -1,10 +1,7 @@
 /**
- * @brief Task interface.
- *
- * The interface of a task that is being executed in self context.
- *
+ * @file      api.Task.hpp
  * @author    Sergey Baigudin, sergey@baigudin.software
- * @copyright 2016-2020, Sergey Baigudin, Baigudin Software
+ * @copyright 2016-2021, Sergey Baigudin, Baigudin Software
  */
 #ifndef API_TASK_HPP_
 #define API_TASK_HPP_
@@ -15,7 +12,13 @@ namespace eoos
 {
 namespace api
 {
-        
+    
+/**
+ * @class Task
+ * @brief Task interface.
+ *
+ * The interface of a task that is being executed in self context.
+ */        
 class Task : public Object
 {
 
@@ -29,7 +32,7 @@ public:
     /**
      * @brief Starts executing an operating system prorgam in itsself context.
      *
-     * @return zero, or error code if an error has been occurred.
+     * @return Zero, or error code if an error has been occurred.
      */
     virtual int32_t start() = 0;
 
@@ -38,9 +41,9 @@ public:
      *
      * The method returns size of stack in bytes which should be allocated for the task.
      *
-     * @return stack size in bytes, or zero if OS default stack size is needed.
+     * @return Stack size in bytes, or zero if OS default stack size is needed.
      */
-    virtual int32_t getStackSize() const = 0;
+    virtual size_t getStackSize() const = 0;
 
 };
